\name{OutlierGrubbs}
\alias{outliergrubbs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
OutlierGrubbs
}
\description{
Removes all outliers from the dataset based on 0.05 significancy by Grubbs-test
}
\usage{
OutlierGrubbs(df)
}
Removes all outliers
\arguments{
  \item{df}{
Dataset
}
}
\details{
Removes all outliers from the dataset based on 0.05 significancy by Grubbs-test
}
\author{
Horvath Mate
}
